Fix CropTransformation for use case with width = 0 and height = 0
As Glide documentation says, Transformations are meant to be statless.
The transform method instead changes the properties width and height
in the very first lines: each property is immutable only if its initial
value is not zero.
Moreover the bitmap toTransform is the original bitmap, so you should
not use its size as the desired size of the resulting bitmap; you have
to use the parameters outWidth and outHeight, instead.